backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.12.5
pydantic-settings==2.1.0
python-binance==1.0.19
supabase==2.28.0
firebase-admin==6.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
websockets>=16.0
httpx>=0.28.1
